Consolidate Correctors and base class.
