--- nimare/correct/__init__.py
@@ -1,14 +1,0 @@
-"""
-Multiple comparisons correction
-"""
-import warnings
-
-from .correct import FDRCorrector, FWECorrector
-
-
-__all__ = [
-    'FDRCorrector',
-    'FWECorrector'
-]
-
-warnings.simplefilter('default')
